api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic-settings==2.7.1
websockets==14.2
python-dotenv==1.0.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
tenacity==9.0.0
langfuse==2.59.3
langgraph>=0.3.0
langgraph-checkpoint-postgres>=2.0.0
psycopg[binary]>=3.1
pgvector>=0.3.0
voyageai>=0.3.0
